fix: lib compilation problem fixed
